+++ tests/test_models.py
@@ -1,18 +1,19 @@
 import datetime
 import logging
 
+from apscheduler import events
 from apscheduler.events import JobExecutionEvent, JobSubmissionEvent
 from django.utils import timezone
 
-from django_apscheduler.jobstores import register_events
 from django_apscheduler.models import DjangoJobExecution
 
 logging.basicConfig()
 
 
 class TestDjangoJobExecutionManager:
-    def test_delete_old_job_executions_deletes_old_jobs(self, db, job, scheduler):
-        register_events(scheduler)
+    def test_delete_old_job_executions_deletes_old_jobs(
+        self, db, djangojobstore, scheduler, job
+    ):
         scheduler.add_job(job, trigger="interval", seconds=1, id="job_1")
         scheduler.add_job(job, trigger="interval", seconds=1, id="job_2")
 
@@ -22,15 +23,25 @@ class TestDjangoJobExecutionManager:
         one_second_ago = now - datetime.timedelta(seconds=1)  # Simulate
 
         scheduler._dispatch_event(
-            JobSubmissionEvent(32768, "job_1", None, [one_second_ago])
+            JobSubmissionEvent(
+                events.EVENT_JOB_SUBMITTED, "job_1", djangojobstore, [one_second_ago],
+            )
         )
 
         scheduler._dispatch_event(
-            JobExecutionEvent(4096, "job_1", None, one_second_ago)
+            JobExecutionEvent(
+                events.EVENT_JOB_EXECUTED, "job_1", djangojobstore, one_second_ago
+            )
         )
 
-        scheduler._dispatch_event(JobSubmissionEvent(32768, "job_2", None, [now]))
-        scheduler._dispatch_event(JobExecutionEvent(4096, "job_2", None, now))
+        scheduler._dispatch_event(
+            JobSubmissionEvent(
+                events.EVENT_JOB_SUBMITTED, "job_2", djangojobstore, [now]
+            )
+        )
+        scheduler._dispatch_event(
+            JobExecutionEvent(events.EVENT_JOB_EXECUTED, "job_2", djangojobstore, now)
+        )
 
         assert DjangoJobExecution.objects.count() == 2